[Security] fixes token refresh on roles changes
