+++ src/main/authentication/Resources/config/services.yml
@@ -6,5 +6,6 @@ imports:
     - { resource: services/manager.yml }
     - { resource: services/security.yml }
     - { resource: services/serializer.yml }
+    - { resource: services/subscriber.yml }
     - { resource: services/updater.yml }
     - { resource: services/voter.yml }
